refactor: ensure all internals are np.float64 and are casted to float() as last as possible (#32)

+++ algorithms/bandits/implementations/action_value_update/AverageSampling.py
@@ -15,7 +15,7 @@ class AverageSampling(ActionValueUpdateStrategy):
     # ==============
 
     def update_action_value(
-        self, q_values: np.ndarray, action: int, reward: float
+        self, q_values: np.ndarray, action: int, reward: np.float64
     ) -> None:
         self._action_counts[action] += 1
         q_values[action] += (1 / self._action_counts[action]) * (
